FFM-11348 Rename retryable request to api
